requirements.txt
llama-cpp-python~=0.2.6
python-dotenv~=1.0.0
pydantic~=1.10.12
uvicorn~=0.23.2
fastapi~=0.99.1
gradio~=3.44.4
chromadb~=0.4.12
langchain~=0.0.297
pymilvus~=2.3.0
socksio~=1.0.0
jq~=1.6.0
sentence-transformers~=2.2.2
torch~=2.0.1